+++ tests/integration_tests/test_subtensor_integration.py
@@ -30,6 +30,7 @@ from tests.helpers import (
     get_mock_keypair,
     get_mock_wallet,
 )
+from bittensor.core.extrinsics import transfer
 
 
 class TestSubtensor(unittest.TestCase):
@@ -123,8 +124,7 @@ class TestSubtensor(unittest.TestCase):
     def test_transfer(self):
         fake_coldkey = get_mock_coldkey(1)
 
-        self.subtensor.do_transfer = MagicMock(return_value=(True, "0x", None))
-        self.subtensor.register = MagicMock(return_value=True)
+        transfer.do_transfer = MagicMock(return_value=(True, "0x", None))
         self.subtensor.get_neuron_for_pubkey_and_subnet = MagicMock(
             return_value=self.mock_neuron
         )
@@ -138,8 +138,7 @@ class TestSubtensor(unittest.TestCase):
 
     def test_transfer_inclusion(self):
         fake_coldkey = get_mock_coldkey(1)
-        self.subtensor.do_transfer = MagicMock(return_value=(True, "0x", None))
-        self.subtensor.register = MagicMock(return_value=True)
+        transfer.do_transfer = MagicMock(return_value=(True, "0x", None))
         self.subtensor.get_neuron_for_pubkey_and_subnet = MagicMock(
             return_value=self.mock_neuron
         )
@@ -152,7 +151,7 @@ class TestSubtensor(unittest.TestCase):
 
     def test_transfer_failed(self):
         fake_coldkey = get_mock_coldkey(1)
-        self.subtensor.do_transfer = MagicMock(
+        transfer.do_transfer = MagicMock(
             return_value=(False, None, "Mock failure message")
         )
 
@@ -176,7 +175,6 @@ class TestSubtensor(unittest.TestCase):
         fake_coldkey = get_mock_coldkey(1)
         self.subtensor.do_transfer = MagicMock(return_value=(True, "0x", None))
 
-        self.subtensor.register = MagicMock(return_value=True)
         self.subtensor.get_neuron_for_pubkey_and_subnet = MagicMock(
             return_value=self.mock_neuron
         )
